requirements.txt
langchain
streamlit
langgraph
python-dotenv
langchain-openai
python-docx
python-pptx
langchain-core
langchain-community
bs4
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
pymupdf
langchain_huggingface
langchain-groq
pytube
youtube-transcript-api
arxiv
uvicorn
fastapi